+++ test/unit/methods/loadData.test.ts
@@ -1,4 +1,4 @@
-import { assertEquals } from "@std/assert";
+import { assertEquals, assertRejects } from "@std/assert";
 import SimpleDB from "../../../src/class/SimpleDB.ts";
 import SimpleTable from "../../../src/class/SimpleTable.ts";
 
@@ -663,5 +663,142 @@ Deno.test("should load data from a specific sheet in an xlsx file and all as tex
     { A1: "Ontario", B1: "15.0" },
     { A1: "British Columbia", B1: "5.0" },
   ]);
+  await sdb.done();
+});
+
+Deno.test("should load only specific columns from a CSV file", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+  await table.loadData(["test/data/files/data.csv"], {
+    columns: ["key1"],
+  });
+
+  const data = await table.getData();
+
+  assertEquals(data, [
+    { key1: "1" },
+    { key1: "3" },
+    { key1: "8" },
+    { key1: "brioche" },
+  ]);
+  await sdb.done();
+});
+
+Deno.test("should load multiple specific columns from a CSV file", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+  await table.loadData(["test/data/files/data.csv"], {
+    columns: ["key2", "key1"],
+  });
+
+  const data = await table.getData();
+
+  assertEquals(data, [
+    { key2: "2", key1: "1" },
+    { key2: "coucou", key1: "3" },
+    { key2: "10", key1: "8" },
+    { key2: "croissant", key1: "brioche" },
+  ]);
+  await sdb.done();
+});
+
+Deno.test("should load only specific columns from a JSON file", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+  await table.loadData(["test/data/files/data.json"], {
+    columns: ["key2"],
+  });
+
+  const data = await table.getData();
+
+  assertEquals(data, [
+    { key2: "un" },
+    { key2: "deux" },
+    { key2: "trois" },
+    { key2: "quatre" },
+  ]);
+  await sdb.done();
+});
+
+Deno.test("should load only specific columns from a Parquet file", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+  await table.loadData(["test/data/files/data.parquet"], {
+    columns: ["key1"],
+  });
+
+  const data = await table.getData();
+
+  assertEquals(data, [
+    { key1: 1 },
+    { key1: 3 },
+    { key1: 8 },
+    { key1: 3 },
+  ]);
+  await sdb.done();
+});
+
+Deno.test("empty columns array should load all columns (CSV)", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+  await table.loadData(["test/data/files/data.csv"], {
+    columns: [],
+  });
+
+  const data = await table.getData();
+
+  assertEquals(data, [
+    { key1: "1", key2: "2" },
+    { key1: "3", key2: "coucou" },
+    { key1: "8", key2: "10" },
+    { key1: "brioche", key2: "croissant" },
+  ]);
+  await sdb.done();
+});
+
+Deno.test("invalid column name should throw a DuckDB error", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+
+  await assertRejects(
+    () =>
+      table.loadData(["test/data/files/data.csv"], {
+        columns: ["nonexistent_column"],
+      }),
+  );
+
+  await sdb.done();
+});
+
+Deno.test("combining columns with Excel files should throw an error", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+
+  await assertRejects(
+    () =>
+      table.loadData(["test/data/files/populations-one-sheet.xlsx"], {
+        columns: ["Country"],
+      }),
+    Error,
+    "columns",
+  );
+
+  await sdb.done();
+});
+
+Deno.test("combining columns with Excel files (by fileType) should throw an error", async () => {
+  const sdb = new SimpleDB();
+  const table = sdb.newTable();
+
+  await assertRejects(
+    () =>
+      table.loadData(["test/data/files/populations-one-sheet.xlsx"], {
+        fileType: "excel",
+        columns: ["Country"],
+      }),
+    Error,
+    "columns",
+  );
+
   await sdb.done();
 });
